Moved all widgets to components dir
