+++ lib/components/group_card_view.dart
