+++ lib/components/group_chat_widget.dart
@@ -1,8 +1,8 @@
-import 'package:connectuni/model/group_card_view.dart';
+import 'package:connectuni/components/group_card_view.dart';
 import 'package:connectuni/screens/group_chat_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'group.dart';
+import '../model/group.dart';
 
 /// GroupChatWidget is a widget that displays the group chat.
 /// It is a clickable widget that takes the user to the group chat page.
