+++ lib/screens/groups_screen.dart
@@ -1,8 +1,8 @@
-import 'package:connectuni/model/group_chat_widget.dart';
+import 'package:connectuni/components/group_chat_widget.dart';
 import 'package:flutter/material.dart';
 
 import 'package:connectuni/model/group.dart';
-import 'package:connectuni/model/group_card_view.dart';
+import 'package:connectuni/components/group_card_view.dart';
 import '/screens/chatpage.dart';
 
 class GroupsScreen extends StatefulWidget {
