+++ lib/screens/search_groups_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
 
 import '../model/group.dart';
-import '../model/group_card_view.dart';
+import '../components/group_card_view.dart';
 
 
 class SearchGroupsScreen extends StatefulWidget {
