+++ lib/screens/search_people_screen.dart
@@ -1,5 +1,5 @@
 import 'package:connectuni/model/userList.dart';
-import 'package:connectuni/model/user_card_search.dart';
+import 'package:connectuni/components/user_card_search.dart';
 import 'package:flutter/material.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
